+++ tgbotapi.api/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/api/LiveLocationProvider.kt
@@ -3,8 +3,8 @@ package dev.inmo.tgbotapi.extensions.api
 import com.soywiz.klock.DateTime
 import com.soywiz.klock.TimeSpan
 import dev.inmo.tgbotapi.bot.TelegramBot
-import dev.inmo.tgbotapi.extensions.api.edit.LiveLocation.editLiveLocation
-import dev.inmo.tgbotapi.extensions.api.edit.LiveLocation.stopLiveLocation
+import dev.inmo.tgbotapi.extensions.api.edit.location.live.editLiveLocation
+import dev.inmo.tgbotapi.extensions.api.edit.location.live.stopLiveLocation
 import dev.inmo.tgbotapi.requests.send.SendLiveLocation
 import dev.inmo.tgbotapi.types.*
 import dev.inmo.tgbotapi.types.buttons.InlineKeyboardMarkup