+++ tgbotapi.api/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/api/edit/location/live/EditInlineMessageLiveLocation.kt
@@ -1,13 +1,11 @@
-package dev.inmo.tgbotapi.extensions.api.edit.LiveLocation
+package dev.inmo.tgbotapi.extensions.api.edit.location.live
 
 import dev.inmo.tgbotapi.bot.TelegramBot
-import dev.inmo.tgbotapi.extensions.api.edit.location.live.editLiveLocation
 import dev.inmo.tgbotapi.requests.edit.location.live.EditInlineMessageLiveLocation
 import dev.inmo.tgbotapi.types.*
 import dev.inmo.tgbotapi.types.buttons.InlineKeyboardMarkup
 import dev.inmo.tgbotapi.types.location.LiveLocation
 
-@Deprecated("Replaced", ReplaceWith("editLiveLocation", "dev.inmo.tgbotapi.extensions.api.edit.location.live.editLiveLocation"))
 suspend fun TelegramBot.editLiveLocation(
     inlineMessageId: InlineMessageIdentifier,
     latitude: Double,
@@ -16,10 +14,13 @@ suspend fun TelegramBot.editLiveLocation(
     heading: Degrees? = null,
     proximityAlertRadius: Meters? = null,
     replyMarkup: InlineKeyboardMarkup? = null
-) = editLiveLocation(inlineMessageId, latitude, longitude, horizontalAccuracy, heading, proximityAlertRadius, replyMarkup)
-@Deprecated("Replaced", ReplaceWith("editLiveLocation", "dev.inmo.tgbotapi.extensions.api.edit.location.live.editLiveLocation"))
+) = execute(
+    EditInlineMessageLiveLocation(
+        inlineMessageId, latitude, longitude, horizontalAccuracy, heading, proximityAlertRadius, replyMarkup
+    )
+)
 suspend fun TelegramBot.editLiveLocation(
     inlineMessageId: InlineMessageIdentifier,
     location: LiveLocation,
     replyMarkup: InlineKeyboardMarkup? = null
-) = editLiveLocation(inlineMessageId, location, replyMarkup)
+) = editLiveLocation(inlineMessageId, location.latitude, location.longitude, location.horizontalAccuracy, location.heading, location.proximityAlertRadius, replyMarkup)