+++ tgbotapi.api/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/api/edit/location/live/StopChatMessageLiveLocation.kt
@@ -1,7 +1,6 @@
-package dev.inmo.tgbotapi.extensions.api.edit.LiveLocation
+package dev.inmo.tgbotapi.extensions.api.edit.location.live
 
 import dev.inmo.tgbotapi.bot.TelegramBot
-import dev.inmo.tgbotapi.extensions.api.edit.location.live.stopLiveLocation
 import dev.inmo.tgbotapi.requests.edit.location.live.StopChatMessageLiveLocation
 import dev.inmo.tgbotapi.types.ChatIdentifier
 import dev.inmo.tgbotapi.types.MessageIdentifier
@@ -14,30 +13,31 @@ import dev.inmo.tgbotapi.types.message.content.LocationContent
  * @param replyMarkup Some [InlineKeyboardMarkup]. See [dev.inmo.tgbotapi.extensions.utils.types.buttons.inlineKeyboard]
  * as a builder for that
  */
-@Deprecated("Replaced", ReplaceWith("stopLiveLocation", "dev.inmo.tgbotapi.extensions.api.edit.location.live.stopLiveLocation"))
 suspend fun TelegramBot.stopLiveLocation(
     chatId: ChatIdentifier,
     messageId: MessageIdentifier,
     replyMarkup: InlineKeyboardMarkup? = null
-) = stopLiveLocation(chatId, messageId, replyMarkup)
+) = execute(
+    StopChatMessageLiveLocation(
+        chatId, messageId, replyMarkup
+    )
+)
 
 /**
  * @param replyMarkup Some [InlineKeyboardMarkup]. See [dev.inmo.tgbotapi.extensions.utils.types.buttons.inlineKeyboard]
  * as a builder for that
  */
-@Deprecated("Replaced", ReplaceWith("stopLiveLocation", "dev.inmo.tgbotapi.extensions.api.edit.location.live.stopLiveLocation"))
 suspend fun TelegramBot.stopLiveLocation(
     chat: Chat,
     messageId: MessageIdentifier,
     replyMarkup: InlineKeyboardMarkup? = null
-) = stopLiveLocation(chat, messageId, replyMarkup)
+) = stopLiveLocation(chat.id, messageId, replyMarkup)
 
 /**
  * @param replyMarkup Some [InlineKeyboardMarkup]. See [dev.inmo.tgbotapi.extensions.utils.types.buttons.inlineKeyboard]
  * as a builder for that
  */
-@Deprecated("Replaced", ReplaceWith("stopLiveLocation", "dev.inmo.tgbotapi.extensions.api.edit.location.live.stopLiveLocation"))
 suspend fun TelegramBot.stopLiveLocation(
     message: ContentMessage<LocationContent>,
     replyMarkup: InlineKeyboardMarkup? = null
-) = stopLiveLocation(message, replyMarkup)
+) = stopLiveLocation(message.chat, message.messageId, replyMarkup)