+++ tgbotapi.api/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/api/edit/reply_markup/EditChatMessageReplyMarkup.kt
@@ -1,7 +1,7 @@
-package dev.inmo.tgbotapi.extensions.api.edit.ReplyMarkup
+package dev.inmo.tgbotapi.extensions.api.edit.reply_markup
 
 import dev.inmo.tgbotapi.bot.TelegramBot
-import dev.inmo.tgbotapi.extensions.api.edit.reply_markup.editMessageReplyMarkup
+import dev.inmo.tgbotapi.requests.edit.reply_markup.EditChatMessageReplyMarkup
 import dev.inmo.tgbotapi.types.ChatIdentifier
 import dev.inmo.tgbotapi.types.MessageIdentifier
 import dev.inmo.tgbotapi.types.buttons.InlineKeyboardMarkup
@@ -12,31 +12,30 @@ import dev.inmo.tgbotapi.types.message.abstracts.Message
  * @param replyMarkup Some [InlineKeyboardMarkup]. See [dev.inmo.tgbotapi.extensions.utils.types.buttons.inlineKeyboard]
  * as a builder for that
  */
-@Deprecated("Replaced", ReplaceWith("editMessageReplyMarkup", "dev.inmo.tgbotapi.extensions.api.edit.reply_markup.editMessageReplyMarkup"))
 suspend fun TelegramBot.editMessageReplyMarkup(
     chatId: ChatIdentifier,
     messageId: MessageIdentifier,
     replyMarkup: InlineKeyboardMarkup? = null
-) = editMessageReplyMarkup(chatId, messageId, replyMarkup)
+) = execute(
+    EditChatMessageReplyMarkup(chatId, messageId, replyMarkup)
+)
 
 /**
  * @param replyMarkup Some [InlineKeyboardMarkup]. See [dev.inmo.tgbotapi.extensions.utils.types.buttons.inlineKeyboard]
  * as a builder for that
  */
-@Deprecated("Replaced", ReplaceWith("editMessageReplyMarkup", "dev.inmo.tgbotapi.extensions.api.edit.reply_markup.editMessageReplyMarkup"))
 suspend fun TelegramBot.editMessageReplyMarkup(
     chat: Chat,
     messageId: MessageIdentifier,
     replyMarkup: InlineKeyboardMarkup? = null
-) = editMessageReplyMarkup(chat, messageId, replyMarkup)
+) = editMessageReplyMarkup(chat.id, messageId, replyMarkup)
 
 /**
  * @param replyMarkup Some [InlineKeyboardMarkup]. See [dev.inmo.tgbotapi.extensions.utils.types.buttons.inlineKeyboard]
  * as a builder for that
  */
-@Deprecated("Replaced", ReplaceWith("editMessageReplyMarkup", "dev.inmo.tgbotapi.extensions.api.edit.reply_markup.editMessageReplyMarkup"))
 suspend fun TelegramBot.editMessageReplyMarkup(
     message: Message,
     replyMarkup: InlineKeyboardMarkup? = null
-) = editMessageReplyMarkup(message, replyMarkup)
+) = editMessageReplyMarkup(message.chat.id, message.messageId, replyMarkup)