+++ tgbotapi.api/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/api/edit/reply_markup/EditInlineMessageReplyMarkup.kt
@@ -1,7 +1,6 @@
-package dev.inmo.tgbotapi.extensions.api.edit.ReplyMarkup
+package dev.inmo.tgbotapi.extensions.api.edit.reply_markup
 
 import dev.inmo.tgbotapi.bot.TelegramBot
-import dev.inmo.tgbotapi.extensions.api.edit.reply_markup.editMessageReplyMarkup
 import dev.inmo.tgbotapi.requests.edit.reply_markup.EditInlineMessageReplyMarkup
 import dev.inmo.tgbotapi.types.InlineMessageIdentifier
 import dev.inmo.tgbotapi.types.buttons.InlineKeyboardMarkup
@@ -10,8 +9,7 @@ import dev.inmo.tgbotapi.types.buttons.InlineKeyboardMarkup
  * @param replyMarkup Some [InlineKeyboardMarkup]. See [dev.inmo.tgbotapi.extensions.utils.types.buttons.inlineKeyboard]
  * as a builder for that
  */
-@Deprecated("Replaced", ReplaceWith("editMessageReplyMarkup", "dev.inmo.tgbotapi.extensions.api.edit.reply_markup.editMessageReplyMarkup"))
 suspend fun TelegramBot.editMessageReplyMarkup(
     inlineMessageId: InlineMessageIdentifier,
     replyMarkup: InlineKeyboardMarkup? = null
-) = editMessageReplyMarkup(inlineMessageId, replyMarkup)
+) = execute(EditInlineMessageReplyMarkup(inlineMessageId, replyMarkup))